update the HeroSearchComponent template
The template contains a text box and a list of matching search results.

As the user types in the search box, a keyup event binding calls the
component's search() method with the new search box value.

Note: The $ is a convention that indicates heroes$ is an Observable, not
an array. The AysncPipe (|) subscribes an Observable automatically.